get rid of all / and % with >> and &
Replace division and modulus with bitwise operation equivalents -- divisions are in power of 2 and modulus is just getting the first 32 bits of a number.
Also remembered to delete intbits.

Trying to improve performance so that it can handle n = 10^8 in HackerRank's environment for all test cases while still using a bit array like the problem name suggests you should. This fails a couple of tests because it takes longer than 2 seconds to perform -- and the math is what pushes it over the edge for time.

The thought here is that division takes a lot more time to do, so getting rid of it should increase performance. And just incase the compiler also uses division for doing modulus, I replaced it with a bit wise operator as well.

Also got rid of string.h library in the hope that would help with performance like getting rid of the other libraries did, but there wasn't a noticeable difference.